backend/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
azure-core==1.38.0
azure-cosmos==4.14.5
azure-identity==1.25.1
azure-storage-blob==12.28.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachelib==0.13.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.4.1
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
Deprecated==1.3.1
distro==1.9.0
et_xmlfile==2.0.0
filelock==3.20.3
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Limiter==4.1.1
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.1.0
fuzzywuzzy==0.18.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai==1.60.0
google-generativeai==0.8.6
google_search_results==2.4.2
googleapis-common-protos==1.72.0
greenlet==3.3.1
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==24.1.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.3.5
idna==3.11
importlib_metadata==8.4.0
importlib_resources==6.5.2
iniconfig==2.3.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.7
langchain-community==0.4.1
langchain-core==1.2.7
langchain-google-genai==4.2.0
langchain-openai==1.1.7
langchain-text-splitters==1.1.0
langsmith==0.6.8
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-checkpoint-sqlite==3.0.3
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
Levenshtein==0.27.3
limits==5.6.0
MarkupSafe==3.0.3
mmh3==5.2.0
msal==1.34.0
msal-extensions==1.3.1
msgspec==0.20.0
multidict==6.7.1
numpy==2.3.5
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.16.0
openpyxl==3.1.5
opentelemetry-api>=1.39.0
opentelemetry-sdk>=1.39.0
opentelemetry-proto>=1.39.0
opentelemetry-semantic-conventions>=0.52b0
orjson==3.11.6
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pillow==12.1.0
pinecone>=5.3.1
pluggy==1.6.0
posthog==5.4.0
propcache==0.4.1
proto-plus==1.27.0
protobuf>=5.26.1,<6.0.0
psutil==7.2.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.7
PyMySQL==1.1.2
pyparsing==3.3.2
pypdf==6.6.2
PyPika==0.50.0
pytest==9.0.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
redis==7.1.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.30.0
rsa==4.9.1
sentence-transformers>=2.2.0
serpapi==0.1.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.0
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.5
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
pymongo>=4.6.0
